chore: structuring billing Co-authored-by: ManikantaPotnuru9176 <ManikantaPotnuru9176@users.noreply.github.com>
